# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Учет бюджета
  description: Простой API для ведения личных финансов
  version: 0.1.0
paths:
...
  /transactions/:
    get:
      tags:
      - Операции
      summary: Read Transactions
      description: Вернуть список операций с указанными фильтрами.
      operationId: read_transactions_transactions__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Start date (posted_at)
          title: Date From
        description: Start date (posted_at)
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: End date (posted_at)
          title: Date To
        description: End date (posted_at)
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Category
          title: Category Id
        description: Category
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Limit
          title: Limit
        description: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
                title: Response Read Transactions Transactions  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Операции
      summary: Create Transaction
      description: Создать новую операцию.
      operationId: create_transaction_transactions__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions/import:
    post:
      tags:
      - Операции
      summary: Import Transactions
      description: Импортировать операции из CSV или Excel.
      operationId: import_transactions_transactions_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_transactions_transactions_import_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /transactions/export:
    get:
      tags:
      - Операции
      summary: Export Transactions
      description: Экспортировать операции в CSV.
      operationId: export_transactions_transactions_export_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Start date (posted_at)
          title: Date From
        description: Start date (posted_at)
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: End date (posted_at)
          title: Date To
        description: End date (posted_at)
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Category
          title: Category Id
        description: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions/{tx_id}:
    get:
      tags:
      - Операции
      summary: Read Transaction
      description: Получить одну операцию.
      operationId: read_transaction_transactions__tx_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: tx_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tx Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Операции
      summary: Update Transaction
      description: Обновить операцию.
      operationId: update_transaction_transactions__tx_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: tx_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tx Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Операции
      summary: Delete Transaction
      description: Удалить операцию.
      operationId: delete_transaction_transactions__tx_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: tx_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tx Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_import_transactions_transactions_import_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_import_transactions_transactions_import_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PostingCreate:
      properties:
        amount:
          type: number
          title: Amount
        side:
          type: string
          title: Side
        account_id:
          type: string
          format: uuid
          title: Account Id
        currency_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency Code
      type: object
      required:
      - amount
      - side
      - account_id
      title: PostingCreate
...
    Transaction:
      properties:
        posted_at:
          type: string
          format: date-time
          title: Posted At
        payee:
          anyOf:
          - type: string
          - type: 'null'
          title: Payee
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
      type: object
      required:
      - posted_at
      - id
      - user_id
      title: Transaction
    TransactionCreate:
      properties:
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        payee:
          anyOf:
          - type: string
          - type: 'null'
          title: Payee
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        postings:
          items:
            $ref: '#/components/schemas/PostingCreate'
          type: array
          title: Postings
      type: object
      title: TransactionCreate
    TransactionUpdate:
      properties:
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        payee:
          anyOf:
          - type: string
          - type: 'null'
          title: Payee
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
      type: object
      title: TransactionUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /users/token
...